get a naive test passing
